backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-filter==23.2
django-password-validators==1.7.1
django-phone-field==1.8.1
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-redis==5.4.0
django-templated-mail==1.1.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-base64==2.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
flake8==6.1.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
phonenumbers==8.13.19
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.4
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.8
